run JUnit tests using android-junit-report to get XML output
Jenkins needs some kind of report from the JUnit tests in order to tell
whether the tests succeeded or not.  android-junit-report is a library to do
exactly that.  With this setup, Jenkins should now successfully understand the
status of the JUnit tests, where before it just ran them and ignored the
results.

It turns out that Jenkins was running the JUnit tests all along, but it just
never reported on them. This adds a jar to the test project that makes JUnit
reports that Jenkins can parse, and the report on the results. So now if the
JUnit tests fail, people will be emailed just like build failures.